Extract common logic into status schema class
To reduce duplicated code, this introduces a common
CalculatedStatusRecordSchema, which handles the logic
for removing the 'status' field on write, as well as
calculating it on read.
