+++ backend/compact-connect/lambdas/common-python/cc_common/data_model/schema/privilege.py
@@ -1,16 +1,15 @@
 # ruff: noqa: N801, N815, ARG002  invalid-name unused-argument
-from datetime import date, datetime
 
 from marshmallow import pre_dump, pre_load
 from marshmallow.fields import UUID, Date, String
 from marshmallow.validate import Length, OneOf
 
 from cc_common.config import config
-from cc_common.data_model.schema.base_record import BaseRecordSchema
+from cc_common.data_model.schema.base_record import BaseRecordSchema, CalculatedStatusRecordSchema
 
 
 @BaseRecordSchema.register_schema('privilege')
-class PrivilegeRecordSchema(BaseRecordSchema):
+class PrivilegeRecordSchema(CalculatedStatusRecordSchema):
     """Schema for privilege records in the license data table"""
 
     _record_type = 'privilege'
@@ -23,7 +22,6 @@ class PrivilegeRecordSchema(BaseRecordSchema):
     dateOfRenewal = Date(required=True, allow_none=False)
     dateOfExpiration = Date(required=True, allow_none=False)
     compactTransactionId = String(required=False, allow_none=False)
-    status = String(required=True, allow_none=False, validate=OneOf(['active', 'inactive']))
 
     # Generated fields
     pk = String(required=True, allow_none=False)
@@ -35,29 +33,9 @@ class PrivilegeRecordSchema(BaseRecordSchema):
         in_data['sk'] = f'{in_data['compact']}#PROVIDER#privilege/{in_data['jurisdiction']}#{in_data['dateOfRenewal']}'
         return in_data
 
-    @pre_dump
-    def remove_status_field_if_present(self, in_data, **kwargs):
-        # status is calculated at load time, so we do not write it to the database
-        in_data.pop('status', None)
-        return in_data
-
     @pre_load
     def pre_load_initialization(self, in_data, **kwargs):  # noqa: ARG001 unused-argument
-        in_data = self._calculate_status(in_data)
         return self._set_date_of_renewal(in_data)
-
-    def _calculate_status(self, in_data, **kwargs):
-        # determine if the status is active or inactive by comparing the expiration date to now
-        in_data['status'] = (
-            'active'
-            if (
-                date.fromisoformat(in_data['dateOfExpiration'])
-                > datetime.now(tz=config.expiration_date_resolution_timezone).date()
-            )
-            else 'inactive'
-        )
-
-        return in_data
 
     def _set_date_of_renewal(self, in_data, **kwargs):
         # for backwards compatibility with the old data model
